+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### **Added**
-- add integration tests for modules 
+
+- add integration tests for modules
   - `service-catalog/app-registry`
   - `testing/integration-tests`
   - `storage/opensearch-serverless`
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `storage/fsx-lustre`
   - `database/rds`
   - `compute/aws-batch`
+- added data repository integration for FSX Lustre module
 - added partition support for modules
   - emr-serverless
   - neptune
@@ -31,14 +33,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - opensearch
 - adds workflow specific to changes for `requirements-dev.txt` so all static checks are run
 
-
 ### **Changed**
 
 - fixed `storage/ecr` module integration test
+- fixed `fsx lustre` integration module to add filesystem permissions on fly using an input var
+- fixed the `fsx lustre` module to use the KubectlV29 image to run kubectl commands
 - update CDK in neptune module due to deprecated neptune version
 - update airflow version to 2.8.1
 - updated `get-modules` workflow to only run tests against changed files in `modules/**`
-
 
 ### **Removed**